+++ tests/utils.js
@@ -56,7 +56,7 @@ export function computeStyle(className) {
   return styleCopy
 }
 
-export const generateClassName = rule => `${rule.key}-id`
+export const createGenerateClassName = () => rule => `${rule.key}-id`
 
 // Make sure tests are isolated.
 afterEach(() => {