+++ changelog.md
@@ -2,6 +2,10 @@
 
 - Option `insertionPoint` can now accept a DOM node.
 
+### Breaking changes
+
+- Jss constructor and Jss.setup option `generateClassName` has been removed. A new option is called `createGenerateClassName` which is a factory that returns the old `generateClassName`. We need this to reset counters on SSR for each request.
+
 ## 7.1.7 / 2017-06-15
 
 - Fix CSS flow types